requirements.txt
django
crewai
langchain
langchain-community
langchain-openai   # so we can do from langchain_openai import AzureOpenAIEmbeddings
chromadb
requests
python-dotenv
gunicorn
uvicorn
django
crewai
langchain
langchain-community
langchain-openai
chromadb
requests
python-dotenv
gunicorn
uvicorn
beautifulsoup4
pillow
